pyproject.toml
[project]
name = "vonageassist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.4.1",
    "fastmcp",
    "beautifulsoup4",
    "python-dotenv",
]
